requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
streamlit==1.41.1
sqlalchemy==2.0.36
pymysql==1.1.1
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
openpyxl==3.1.5
requests==2.32.3
yfinance==0.2.50
python-dotenv==1.0.1
